tj3JPEGBufSize(): Fix int overflow w/ large dims
The intermediate PAD(width, mcuw) * PAD(height, mcuh) computation was
performed using signed integer arithmetic and could thus overflow if
PAD(width, mcuw) * PAD(height, mcuh) > INT_MAX.  This was innocuous,
however, because the API requires that the allocated buffer size be
reported to tj3Compress*() or tj3Transform().  The calling program would
presumably pass the (incorrect) return value from tj3JPEGBufSize() to
tj3Alloc() and report the same size to tj3Compress*() or tj3Transform().
Depending on the value of TJPARAM_NOREALLOC,
tj3Compress*()/tj3Transform() would either reallocate the buffer as
needed or fail if the reported buffer size was exceeded.  As long as the
reported buffer size was the same as the allocated buffer size, no
buffer overrun could occur.  Thus, the only issue was a usability issue:
passing large dimensions to tj3JPEGBufSize(), then passing the return
value of tj3JPEGBufSize() to tj3Alloc() and
tj3Compress*()/tj3Transform(), could have caused an error ("Buffer
passed to JPEG library is too small") if TJPARAM_NOREALLOC was set.

+++ src/turbojpeg.c
@@ -946,7 +946,8 @@ DLLEXPORT size_t tj3JPEGBufSize(int width, int height, int jpegSubsamp)
   mcuw = tjMCUWidth[jpegSubsamp];
   mcuh = tjMCUHeight[jpegSubsamp];
   chromasf = jpegSubsamp == TJSAMP_GRAY ? 0 : 4 * 64 / (mcuw * mcuh);
-  retval = PAD(width, mcuw) * PAD(height, mcuh) * (2ULL + chromasf) + 2048ULL;
+  retval = (unsigned long long)PAD(width, mcuw) * PAD(height, mcuh) *
+           (2ULL + chromasf) + 2048ULL;
 #if ULLONG_MAX > ULONG_MAX
   if (retval > (unsigned long long)((unsigned long)-1))
     THROWG("Image is too large", 0);
